requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tenacity>=8.2.0
click>=8.1.0
rich>=13.0.0
PyYAML>=6.0